Fixing problem with definition Typescript of PouchDB
The original definition breaks as soon as the PouchDB object is used. For example to
create a new database(using the new keyword) or to sync a database(PouchDB.sync).

The Typescript compiler would return the following error when used with the 'new' keyword:

TypeScript error: app/app.ts(7,10): Error TS2351: Cannot use 'new' with an expression whose type lacks a call or construct signature.